Add dataset report
This will be important to check that dates fall within the expected range
(i.e. when we are doing weekly extracts, we should not have dates in the 4th
week in the month included in the extract for 2nd week of the month).
The study definition is parameterised, and the default is to use
last_day_of_month(index_date). But when computing measures the
parameters should be ignored.

The report groups dates by year or month, and counts impossible dates
(before the patients birth year), and dates in the future.