Add optional argument for measure to specify first
The default order of the plots is alphabetical using the measure name.
This may not be ideal if an overall/population measure appears after
breakdown plots. As a partial fix, allow the user to specify one plot to
come first. A future fix could be allow the user to specify the entire
order.

Also switch from a rate to a rate per 1000.