Add curation study def for new diagnostic codes
We do not know how/if fever pain and throat swabs will be coded. Create
a study def for one month to assess these codelists.

Change the format to feather so we can be sure to describe the codes. If
the file type is csv or csv.gz, we need a better way to determine a
categorical variable. Or are we okay outputting each level of the
variable with redaction? Or just the top n most common levels?