Explore date range and with variable matches
Here we explore two things:
1. The impact of date_format="YYYY-MM-DD" versus "YYYY-MM", as we think
if the later is used in combination with -7 days + 14 days, it would
always look -7 to +14 days after the first of the month. Compare the
number of matches in the whole month, to -7 + 14 days, to same day with
the full format string, to same day with the short format string

2. Count the number of days between a scarlet fever diagnosis and any
antibiotic prescription and group the patients by the number of days.
Apply redaction and rounding before displaying in the html report.