Add x-tick frequency option to panel plots
The monthly report currently has 73 months. This is too many to report
every x-axis label. Provide option when creating panel plots for how
often to report the x-tick label (the default is every label).

Ideally we would want to always report the first and last tick label,
but in the case where there are 73 months (prime number) there is no way
to include both the first and last tick with an even distribution.