fix: check columns are the same, not the rows
When creating the event counts table, we add a column for the
denominator count. For this report, the columns should be using the same
population, so the denominators should be the same and we only need to
report one column. We were previously incorrectly asserting that the
rows were all the same.