Keep measures for both rate and percentage
The rate of patients with a prescription and clinical indication over
the total population is probably more clinically helpful than the
percentage (over patients with a prescription).

Change the primary measures to be over the population (that is the
measure that will be picked up by the plots and the report), and create
a new measure for the percentage.