Add medication w/ clinical, clinical w/medication
Single plot broken down by medication or clinical but using the "with"
variables.
These plots will likely be more helpful for the publication, and we can
see whether/where we want to add them in the report.

Remove the code individual plots by measure as we are not using them,
and the 'return' statement was skipping them anyway.