Fix table1 order and add confidence intervals
Age bands need to be ordered numerically, rather than alphabetically.
Add confidence intervals for the rate.
Specify at load time the data types of the columns, so we don't have to
cast after the fact.
The code as written does not work on the os image. Move it down to the
bottom of the project yaml and mark as such.