Make table1 runnable on OS image
It would be nicer if we could specify that our count variables are ints
and not floats, so that we don't have to round when we want to display.
But the os numpy sqrt function cannot handle Int32/Int64 with nans. So
we go back to using float instead.

Minor fixes when testing with REDACTED data.