Move group_medications into separate action
For the paper, we want a panel plot, with one panel for each medication
group.

Ideally we would make new measures to count patients with each
medication type, but this has been done after-the-fact.

Panel plots subsetting requires a measures file and measures pattern.
But because the grouped medications measures do not exist in the
measures file we cannot do this.

Make a new action to make a measures file. This way it can be fed
directly into a panel plot action.