Fix panel plot proportion
These time series are easier to read when they are a bit wider. First
actually allow the user to control how many columns are in the plot.

And now that we are using a panel plot for multiple population
level measures, allow the user to hide the legend (and resize
accordingly) if the legend is redundant.