Move season table and season annotation in utils
Depending on what type of plot we want to generate (a panel plot, or a
single plot with plot measures report), we want to be able to make a
season table and annotate the seasons in either case.

Move this code into utils so it can be used by both scripts.
Annotating the plot with max/min arrows and counts was a bit cluttered
and confusing, and less necessary since we are not plotting on the log
scale. Instead just plot a marker for the max, and refer to the season
table.