Count the number excluded by each criteria
Add a new study def so we can count the excluded population.
This could be in one study def (to ensure they get run at the same
time), but that would mean that all variables are extracted for the
excluded population, and all measures would need to add an AND included
to the numerator and denominator.

Add a processing script to count and then redact and round the criteria
of interest. We could also use the measures framework, but this is
faster for now.